refactor: consumer parameters actions
