+++ src/components/@shared/FormInput/InputElement/ConsumerParameters/Actions.module.css
@@ -1,0 +1,5 @@
+.actions {
+  display: flex;
+  justify-content: center;
+  gap: var(--spacer);
+}
