+++ src/components/@shared/FormInput/InputElement/ConsumerParameters/index.module.css
@@ -18,9 +18,3 @@
 .error {
   composes: error from '../../index.module.css';
 }
-
-.actions {
-  display: flex;
-  justify-content: center;
-  gap: var(--spacer);
-}
